Update worging prototype 2024-10-08

+++ JobApplyOrganizer/JobApplyOrganizer/NewJobProject.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement.TaskbarClock;
 
 namespace JobApplyOrganizer
 {
@@ -21,15 +22,15 @@ namespace JobApplyOrganizer
             NEW = 0,
             EDIT = 1,
         }
-
         JobApplication application = new JobApplication();
-        String installpath = "";
+        String _workdirpath = "";
+        String _programLocationPath = "";
         String FOLDERname = "";
         String HTMLname = "";
+        String _date = "";
         List<JobApplication> jobNewList = new List<JobApplication>();
-        String[] returnLoad { get; set; }
         Enum _type = Type.NEW;
-        public NewJobProject(Enum type, String[] payload, JobApplication jobApplication, String installpath)
+        public NewJobProject(Enum type, String[] payload, JobApplication jobApplication, String workdirpath, String programLocationPath)
         {
             InitializeComponent();
             this._type = type;
@@ -37,7 +38,6 @@ namespace JobApplyOrganizer
             {
                 this.Text = "New";
                 buttonCreate.Text = "Create";
-
             }
             else if (type.ToString() == Type.EDIT.ToString())
             {
@@ -50,50 +50,78 @@ namespace JobApplyOrganizer
             textBoxPhone.Text = payload[3];
             textBoxEmail.Text = payload[4];
             textBoxURL.Text = payload[5];
+            _date = payload[6];
+
+            Console.WriteLine("P "+payload[6]+" P");
+
             this.FOLDERname = textJobtitle.Text + "_" + textCompany.Text;
             this.HTMLname = textJobtitle.Text;
-            this.installpath = installpath;
+            this._workdirpath = workdirpath;
+            this._programLocationPath = programLocationPath;
+
+
+            Console.WriteLine(" -+- "+workdirpath + " " + programLocationPath);
+
         }
-        private void buttonCancel_Click(object sender, EventArgs e)
+        private void ButtonCancel_Click(object sender, EventArgs e)
         {
             this.Close();
         }
-        private void buttonCreate_Click(object sender, EventArgs e)
+        private void ButtonCreate_Click(object sender, EventArgs e)
         {
+            MakeHTML pageHtml = new MakeHTML();
+            String[] inDataHtml = new String[9];
             if (_type.ToString() == Type.NEW.ToString())
             {
                 date = DateTime.Now;
-                String jobPath = this.installpath + "\\P" + date.ToString("yyyyMMdd") + "_" + FOLDERname + "\\";
-                String templatePath = this.installpath + "\\Templates\\";
-                String[] fileList = Util.CreateJobPath(templatePath, jobPath);
-
-                SaveKontaktTXT(jobPath);
-                MakeHTML pageHtml = new MakeHTML();
-                String path = pageHtml.CreateHTML(fileList, this.HTMLname, jobPath, templatePath, date);
-                Console.WriteLine(date.ToString("yyyyMMdd ") + installpath);
-                Console.WriteLine(path);
+                // jobPath
+                inDataHtml[0] = this._workdirpath + "\\P" + date.ToString("yyyyMMdd") + "_" + FOLDERname + "\\";
+                // templatePath
+                inDataHtml[1] = this._workdirpath + "\\Templates\\";
+                // html filename
+                inDataHtml[2] = inDataHtml[0] + HTMLname;
+                // pagename
+                inDataHtml[3] = textBoxName.Text;
+                inDataHtml[4] = textBoxPhone.Text;
+                inDataHtml[5] = textBoxEmail.Text;
+                inDataHtml[6] = textBoxURL.Text;
+                inDataHtml[7] = this.HTMLname;
+                inDataHtml[8] = this._programLocationPath;
+                Console.WriteLine("inDataHtml\n{0}\n{1}\n{2}\n{3}\n{4}\n{5}\n{6}\n{7}\n{8}", inDataHtml[0], inDataHtml[1], inDataHtml[2], inDataHtml[3], inDataHtml[4], inDataHtml[5], inDataHtml[6], inDataHtml[7], inDataHtml[8]);
+             
+                pageHtml.CreateHTML(inDataHtml);
             }
             else if (_type.ToString() == Type.EDIT.ToString())
             {
-                date = DateTime.Now;
-                String jobPath = this.installpath + "\\P" + date.ToString("yyyyMMdd") + "_" + FOLDERname + "\\";
-                String templatePath = this.installpath + "\\Templates\\";
-                String[] fileList = Util.CreateJobPath(templatePath, jobPath);
-                SaveKontaktTXT(jobPath);
-                Console.WriteLine(jobPath + HTMLname + "++++++++++++++++++++++++++++++++++++++++++++++++++++++++++");
-                Console.WriteLine("++++++++++++++++++++++++++++++++++++++++++++++++++++++++++");
-                String temp = Util.CleanPath((jobPath + HTMLname), "\\\\", "\\");
-                Console.WriteLine(temp + "++++++++++++++++++++++++++++++++++++++++++++++++++++++++++");
-                temp = temp + ".html";
-                File.Delete(temp);
+                // jobPath
+                inDataHtml[0] = this._workdirpath + "\\P" + _date + "_" + FOLDERname + "\\";
+                // templatePath
+                inDataHtml[1] = this._workdirpath + "\\Templates\\";
+                // html filename
+                inDataHtml[2] = inDataHtml[0] + HTMLname;
+                // pagename
+                inDataHtml[3] = textBoxName.Text;
+                inDataHtml[4] = textBoxPhone.Text;
+                inDataHtml[5] = textBoxEmail.Text;
+                inDataHtml[6] = textBoxURL.Text;
+                inDataHtml[7] = this.HTMLname;
 
-                MakeHTML pageHtml = new MakeHTML();
-                String path = pageHtml.CreateHTML(fileList, this.HTMLname, jobPath, templatePath, date);
-                Console.WriteLine(date.ToString("yyyyMMdd ") + installpath);
-                Console.WriteLine(path);
+                String dateStr = inDataHtml[0];
+                int len = this._workdirpath.Length;
+                //dateStr = dateStr.Replace( , len);
+
+                Console.WriteLine(" KOLLA HÄR "+ FOLDERname + " ???");
+
+                String temp = Util.CleanPath((inDataHtml[0] + HTMLname + ".html"), "\\\\", "\\");
+
+                if (File.Exists(temp))
+                {
+                    File.Delete(temp);
+                }
+                pageHtml.CreateHTML(inDataHtml);
             }
 
-
+            this.Close();
         }
         private void SaveKontaktTXT(String path)
         {
